[script] Skip brigtnes restore if lightsensor enabled on openauto start

+++ stage2/03-crankshaft-base/files/opt/crankshaft/service_openauto.sh
@@ -13,7 +13,7 @@ if [ -f /tmp/start_openauto ]; then
     /usr/local/bin/crankshaft display on
 
     # check for day/night on startup
-    if [ $RTC_DAYNIGHT -eq 1 ]; then
+    if [ $RTC_DAYNIGHT -eq 1 ] && [ ! -f /etc/cs_lightsensor ]; then
         if [ $(date +%H) -gt $((RTC_DAY_START-1)) ] && [ $(date +%H) -lt $RTC_NIGHT_START ]; then
             sudo rm -f /tmp/night_mode_enabled > /dev/null 2>&1
             /usr/local/bin/crankshaft brightness restore &
